+++ examples/1d_stencil/1d_stencil_4.cpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2014-2022 Hartmut Kaiser
+//  Copyright (c) 2014-2025 Hartmut Kaiser
 //  Copyright (c) 2014 Patricia Grubel
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -66,8 +66,8 @@ public:
       : data_(new double[size])
       , size_(size)
     {
-        double base_value = double(initial_value * size);
-        for (std::size_t i = 0; i != size; ++i)
+        double base_value = initial_value * double(size);
+        for (std::ptrdiff_t i = 0; i != static_cast<std::ptrdiff_t>(size); ++i)
             data_[i] = base_value + double(i);
     }
 
@@ -186,13 +186,13 @@ struct stepper
             {
                 next[0].then([sem, t](partition&&) {
                     // inform semaphore about new lower limit
-                    sem->signal(t);
+                    sem->signal(static_cast<std::int64_t>(t));
                 });
             }
 
             // suspend if the tree has become too deep, the continuation above
             // will resume this thread once the computation has caught up
-            sem->wait(t);
+            sem->wait(static_cast<std::int64_t>(t));
         }
 
         // Return the solution at time-step 'nt'.
