+++ examples/transpose/transpose_block.cpp
@@ -215,7 +215,7 @@ int hpx_main(hpx::program_options::variables_map& vm)
         verbose = vm.count("verbose") ? true : false;
 
         std::uint64_t bytes =
-            static_cast<std::uint64_t>(2.0 * sizeof(double) * order * order);
+            static_cast<std::uint64_t>(2 * sizeof(double) * order * order);
 
         std::uint64_t num_blocks = num_localities * num_local_blocks;
 
@@ -275,7 +275,8 @@ int hpx_main(hpx::program_options::variables_map& vm)
                 {
                     double col_val =
                         COL_SHIFT * static_cast<double>(b * block_order + j);
-                    A_ptr->data_[i * block_order + j] = col_val + ROW_SHIFT * i;
+                    A_ptr->data_[i * block_order + j] =
+                        col_val + ROW_SHIFT * static_cast<double>(i);
                     B_ptr->data_[i * block_order + j] = -1.0;
                 }
             }
@@ -362,7 +363,9 @@ int hpx_main(hpx::program_options::variables_map& vm)
                 avgtime = avgtime /
                     static_cast<double>((std::max) (iterations - 1,
                         static_cast<std::uint64_t>(1)));
-                std::cout << "Rate (MB/s): " << 1.e-6 * bytes / mintime << ", "
+                std::cout << "Rate (MB/s): "
+                          << 1.e-6 * static_cast<double>(bytes) / mintime
+                          << ", "
                           << "Avg time (s): " << avgtime << ", "
                           << "Min time (s): " << mintime << ", "
                           << "Max time (s): " << maxtime << "\n";
@@ -478,7 +481,7 @@ double test_results(std::uint64_t order, std::uint64_t block_order,
             double errsq = 0.0;
             for (std::uint64_t i = 0; i < order; ++i)
             {
-                double col_val = COL_SHIFT * i;
+                double col_val = COL_SHIFT * static_cast<double>(i);
                 for (std::uint64_t j = 0; j < block_order; ++j)
                 {
                     double diff = trans_block[i * block_order + j] -
